Refactor utils in client sdk integration test
This allows to reuse the setup functionality for other tests using the
integration framework. This refactoring is required for the new hello
world tutorial for the client SDK.
